+++ mempool/src/error/ban_score.rs
@@ -118,6 +118,7 @@ impl MempoolBanScore for ChainstateError {
             ChainstateError::FailedToReadProperty(_) => 0,
             ChainstateError::BootstrapError(_) => 0,
             ChainstateError::BlockInvalidatorError(_) => 0,
+            ChainstateError::IoError(_) => 0,
         }
     }
 }